fixup: rename and move testUtils.tsx
